app/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile app/requirements.in
#
asgiref==3.4.1
    # via django
cffi==1.14.6
    # via cryptography
cryptography==3.4.8
    # via django-rest-knox
django==3.2.7
    # via
    #   -r app/requirements.in
    #   django-filter
    #   django-rest-knox
    #   djangorestframework
django-filter==2.4.0
    # via -r app/requirements.in
django-rest-knox==4.1.0
    # via -r app/requirements.in
django-rest-passwordreset==1.2.0
    # via -r app/requirements.in
djangorestframework==3.12.4
    # via
    #   -r app/requirements.in
    #   django-rest-knox
djangorestframework-camel-case==1.2.0
    # via -r app/requirements.in
pycparser==2.20
    # via cffi
pytz==2021.1
    # via django
sqlparse==0.4.2
    # via django
